+++ src/oma.cpp
@@ -24,7 +24,7 @@ using std::string;
 using std::vector;
 using std::unique_ptr;
 
-TOma::TOma(const string& filename, const string&, uint8_t /*numChannel*/,
+TOma::TOma(const string& filename, const string&, size_t /*numChannel*/,
     uint32_t /*numFrames*/, int cid, uint32_t framesize, bool jointStereo) {
     oma_info_t info;
     info.codec = cid;
@@ -52,6 +52,6 @@ string TOma::GetName() const {
     return {};
 }
 
-uint8_t TOma::GetChannelNum() const {
+size_t TOma::GetChannelNum() const {
     return 2; //for ATRAC3
 }
