+++ src/oma.h
@@ -25,10 +25,10 @@
 class TOma : public ICompressedOutput {
     OMAFILE* File;
 public:
-    TOma(const std::string& filename, const std::string& title, uint8_t numChannel,
+    TOma(const std::string& filename, const std::string& title, size_t numChannel,
         uint32_t numFrames, int cid, uint32_t framesize, bool jointStereo);
     ~TOma();
     void WriteFrame(std::vector<char> data) override;
     std::string GetName() const override;
-    uint8_t GetChannelNum() const override;
+    size_t GetChannelNum() const override;
 };
